requirements.txt
requests
python-dotenv
pylatexenc
nougat-ocr